pyproject.toml
[tool.poetry]
name = "notion_toolkit"
version = "0.0.0"
description = "Notion tookkit: Implement help function for notion in Python"
authors = ["HappyPydog <happydog0930@gmial.com>"]
readme = "README.md"
packages = [{ include = "notion_toolkit", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.6.3"
rich = "^13.7.1"
python-dotenv = "^1.0.1"
notion-client = "^2.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
black = {extras = ["jupyter"], version = "^24.2.0"}
isort = "^5.13.2"
flake8 = "^7.0.0"
ipykernel = "^6.29.3"


[tool.poetry.group.cli.dependencies]
click = "^8.1.7"
typer = "^0.9.0"
fire = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
